# contracts/control-api.openapi.json
{
  "components": {
    "schemas": {
      "ActionAcceptanceView": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "executed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "state",
          "executed_at"
        ],
        "title": "ActionAcceptanceView",
        "type": "object"
      },
      "ActionChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "expected_state_hash": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Expected State Hash",
            "type": "string"
          },
          "expected_state_version": {
            "minimum": 0.0,
            "title": "Expected State Version",
            "type": "integer"
          },
          "expires_in": {
            "default": 30,
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Expires In",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "interrupt",
              "approve",
              "continue_once",
              "inject",
              "publish_repair",
              "cancel_repair",
              "retry_repair"
            ],
            "title": "Kind",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "target": {
            "$ref": "#/components/schemas/ActionTargetRequest"
          }
        },
        "required": [
          "target",
          "device_id",
          "kind",
          "expected_state_version",
          "expected_state_hash"
        ],
        "title": "ActionChallengeRequest",
        "type": "object"
      },
      "ActionChallengeView": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "canonical_payload": {
            "title": "Canonical Payload",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "state",
          "nonce",
          "issued_at",
          "expires_at",
          "canonical_payload"
        ],
        "title": "ActionChallengeView",
        "type": "object"
      },
      "ActionCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ActionCollection",
        "type": "object"
      },
      "ActionTargetRequest": {
        "additionalProperties": false,
        "properties": {
          "host_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Host Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "session",
              "repair",
              "host",
              "verification",
              "repository"
            ],
            "title": "Kind",
            "type": "string"
          },
          "target_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "host_id"
        ],
        "title": "ActionTargetRequest",
        "type": "object"
      },
      "ActionTargetView": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id"
        ],
        "title": "ActionTargetView",
        "type": "object"
      },
      "ActionView": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "canonical_payload": {
            "title": "Canonical Payload",
            "type": "string"
          },
          "effect": {
            "title": "Effect",
            "type": "string"
          },
          "executed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          },
          "expected_state": {
            "title": "Expected State",
            "type": "string"
          },
          "expected_state_hash": {
            "title": "Expected State Hash",
            "type": "string"
          },
          "expected_state_version": {
            "title": "Expected State Version",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "host_available": {
            "title": "Host Available",
            "type": "boolean"
          },
          "host_id": {
            "title": "Host Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          },
          "parameters_hash": {
            "title": "Parameters Hash",
            "type": "string"
          },
          "requires_biometric": {
            "title": "Requires Biometric",
            "type": "boolean"
          },
          "risk": {
            "enum": [
              "medium",
              "high"
            ],
            "title": "Risk",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/ActionTargetView"
          },
          "target_label": {
            "title": "Target Label",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "target",
          "target_label",
          "host_id",
          "kind",
          "state",
          "effect",
          "risk",
          "requires_biometric",
          "parameters_hash",
          "expected_state",
          "expected_state_version",
          "expected_state_hash",
          "nonce",
          "canonical_payload",
          "host_available",
          "issued_at",
          "expires_at",
          "executed_at"
        ],
        "title": "ActionView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignedActionAcceptance": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Action Id",
            "type": "string"
          },
          "device_algorithm": {
            "enum": [
              "Ed25519",
              "P-256"
            ],
            "title": "Device Algorithm",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "device_key_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Device Key Id",
            "type": "string"
          },
          "device_signature": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Device Signature",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "device_id",
          "device_key_id",
          "device_algorithm",
          "device_signature"
        ],
        "title": "SignedActionAcceptance",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned LoopGuard control-plane contract for web and native clients.",
    "title": "LoopGuard Control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/actions": {
      "get": {
        "operationId": "list_actions_v1_actions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionCollection"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Actions",
        "tags": [
          "actions"
        ]
      },
      "post": {
        "operationId": "accept_signed_action_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedActionAcceptance"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionAcceptanceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Signed Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/actions/challenge": {
      "post": {
        "operationId": "create_action_challenge_v1_actions_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionChallengeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Action Challenge",
        "tags": [
          "actions"
        ]
      }
    },
    "/v1/actions/{action_id}": {
      "get": {
        "operationId": "read_action_v1_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Action",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/v1/sessions": {
      "get": {
        "operationId": "list_sessions_v1_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Sessions V1 Sessions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/v1/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Session V1 Sessions  Session Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted API",
      "url": "https://api.loopguard.example"
    }
  ],
...
}